# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  controller.RechargePackageCreateRequest:
    properties:
      amount:
        type: number
      description:
        type: string
      gift_amount:
        type: number
      gift_points:
        type: integer
      name:
        type: string
    type: object
  controller.RechargePackageUpdateRequest:
    properties:
      amount:
        type: number
      description:
        type: string
      gift_amount:
        type: number
      gift_points:
        type: integer
      name:
        type: string
      status:
        type: integer
    type: object
  controller.RechargeRecordCreateRequest:
    properties:
      amount:
        type: number
      gift_amount:
        type: number
      payment_id:
        type: integer
      status:
        type: integer
      user_id:
        type: integer
    required:
    - amount
    - user_id
    type: object
...
  model.RechargePackage:
    properties:
      amount:
        type: number
      created_at:
        type: string
      description:
        type: string
      gift_amount:
        type: number
      gift_points:
        type: integer
      id:
        type: integer
      name:
        type: string
      sort_order:
        type: integer
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        description: 响应状态码
        type: integer
      data:
        description: 响应数据，omitempty表示为空时不序列化
      message:
        description: 响应消息
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 棋室管理系统后端 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 棋室管理系统 API
  version: "1.0"
paths:
...
  /payments:
    get:
      consumes:
      - application/json
      description: 获取所有支付记录列表
      parameters:
      - description: 用户ID
        in: query
        name: user_id
        type: string
      - description: 支付类型
        in: query
        name: payment_type
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取支付列表
      tags:
      - 支付管理
  /payments/{id}:
    delete:
      consumes:
      - application/json
      description: 删除支付记录
      parameters:
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除支付记录
      tags:
      - 支付管理
    get:
      consumes:
      - application/json
      description: 根据ID获取支付记录详情
      parameters:
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取支付详情
      tags:
      - 支付管理
    put:
      consumes:
      - application/json
      description: 更新支付记录信息
      parameters:
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      - description: 支付信息
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新支付记录
      tags:
      - 支付管理
  /recharge-packages:
    get:
      consumes:
      - application/json
      description: 获取储值套餐列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.RechargePackage'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取储值套餐列表
      tags:
      - 储值套餐
    post:
      consumes:
      - application/json
      description: 管理员创建储值套餐
      parameters:
      - description: 储值套餐信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.RechargePackageCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.RechargePackage'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建储值套餐
      tags:
      - 储值套餐
  /recharge-packages/{id}:
    delete:
      consumes:
      - application/json
      description: 管理员删除储值套餐
      parameters:
      - description: 套餐ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除储值套餐
      tags:
      - 储值套餐
    get:
      consumes:
      - application/json
      description: 根据套餐ID获取储值套餐详情
      parameters:
      - description: 套餐ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.RechargePackage'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取储值套餐详情
      tags:
      - 储值套餐
    put:
      consumes:
      - application/json
      description: 管理员更新储值套餐信息
      parameters:
      - description: 套餐ID
        in: path
        name: id
        required: true
        type: string
      - description: 储值套餐更新信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.RechargePackageUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.RechargePackage'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新储值套餐
      tags:
      - 储值套餐
  /recharge-records:
    get:
      consumes:
      - application/json
      description: 获取所有充值记录列表
      parameters:
      - description: 用户ID
        in: query
        name: user_id
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取充值记录列表
      tags:
      - 充值记录管理
    post:
      consumes:
      - application/json
      description: 创建新的充值记录
      parameters:
      - description: 充值记录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.RechargeRecordCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建充值记录
      tags:
      - 充值记录管理
  /recharge-records/{id}:
    delete:
      consumes:
      - application/json
      description: 删除充值记录
      parameters:
      - description: 充值记录ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除充值记录
      tags:
      - 充值记录管理
    get:
      consumes:
      - application/json
      description: 根据ID获取充值记录详情
      parameters:
      - description: 充值记录ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取充值记录详情
      tags:
      - 充值记录管理
    put:
      consumes:
      - application/json
      description: 更新充值记录信息
      parameters:
      - description: 充值记录ID
        in: path
        name: id
        required: true
        type: string
      - description: 充值记录信息
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新充值记录
      tags:
      - 充值记录管理
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"